+++ test/generators/resource_generator_test.rb
@@ -8,7 +8,6 @@ class ResourceGeneratorTest < Rails::Generators::TestCase
   destination File.join(Rails.root, "tmp")
   setup :prepare_destination
 
-
   test "create template" do
     run_generator ["post"]
     assert_file "app/assets/javascripts/templates/post.hbs"