+++ test/generators/view_generator_test.rb
@@ -52,4 +52,3 @@ class ViewGeneratorTest < Rails::Generators::TestCase
     end
   end
 end
-