+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.2.0
 
 * Ember-Data setting up - Brian Cardarella
+* Rails resource generator only overrides if --ember option is given - Brian Cardarella